Ensure partial names never contain backslashes
The result of path.join is used as the name when registering partials. Whenever the partial is in a subdirectory, this presents a problem if the site is built on Windows as the name will contain a backslash instead of the expected forward slash. This results in handlebars being unable to find the partial. Exactly the same problem and fix as https://github.com/superwolff/metalsmith-layouts/pull/82. The 'should accept a partials option' test already covers this case; without this fax, the test fails on Windows due to not finding the partial.